+++ index-test.php
@@ -8,19 +8,17 @@ use ModernPHPException\ModernPHPException;
 $config = __DIR__ . '/config.example.yaml';
 //$config = "";
 $exc = new ModernPHPException($config);
-#$exc->enableOccurrences();
+//$exc->enableOccurrences();
 $exc->ignoreErrors([E_USER_DEPRECATED, E_WARNING]);
 $exc->start();
 
 #http_response_code(404);
-
 #echo '<pre>';
 
-//throw new Exception("Error Processing Request");
+//throw new Exception("<script>alert('test from JS')</script>");
 
-//error_reporting(E_ALL & ~E_USER_DEPRECATED);
-trigger_error("Test", E_USER_DEPRECATED);
-echo "After trigger_error";
+trigger_error("Test trigger", E_USER_DEPRECATED);
+echo "After trigger function";
 
 //UserTest::staticCall();
 
@@ -33,24 +31,17 @@ echo "After trigger_error";
 };
 
 try {
-    echo dividir(5,0)."<br/>";
+    echo dividir(5,0);
 } catch (\Exception $e) {
     $exc->exceptionHandler($e);
     //echo $e->getMessage();
 } */
 
-/* try {
-    echo $a;
-} catch (\Exception $e) {
-    $exc->exceptionHandler($e);
-} */
 
 /* $a = (new UserTest())->secondCall();
-
 var_dump_debug($a, true); */
 
-echo $a;
+//echo $a;
 
 //$a = new FakeClass(); //FATAL ERROR
-
 //(new UserTest())->triggerTest();