+++ src/Exception/OccurrencesException.php
@@ -1,0 +1,5 @@
+<?php
+
+namespace ModernPHPException\Exception;
+
+class OccurrencesException extends \Exception {}