+++ src/Solution.php
@@ -25,17 +25,21 @@ class Solution
     private static string $button_name = "";
 
     /**
+     * Name of solution to fix exception
+     * 
      * @param string $title
      * 
-     * @return Solution
+     * @return static
      */
-    public static function createSolution(string $title): Solution
+    public static function createSolution(string $title): static
     {
         self::$solution_title = $title;
         return new static();
     }
 
     /**
+     * Detailed description of exception solution
+     * 
      * @param string $description
      * 
      * @return Solution
@@ -47,6 +51,9 @@ class Solution
     }
 
     /**
+     * If a documentation exists, this method will display a button for a documentation.
+     * By default, the name of the button will be `Read More`, but you can change the name by changing the second parameter of the method
+     * 
      * @param string $docs_link
      * @param string $button_name
      * 
@@ -56,11 +63,12 @@ class Solution
     {
         self::$solution_docs = $docs_link;
         self::$button_name = $button_name;
-
         return $this;
     }
 
     /**
+     * Get solution title
+     * 
      * @return string
      */
     public function getTitle(): string
@@ -69,6 +77,8 @@ class Solution
     }
 
     /**
+     * Get solution description
+     * 
      * @return string
      */
     public function getDescription(): string
@@ -77,6 +87,8 @@ class Solution
     }
 
     /**
+     * Get solution docs
+     * 
      * @return array
      */
     public function getDocs(): array